Removed direct access to CustomerIO that is not used.
