+++ src/main/kotlin/com/hedvig/underwriter/serviceIntegration/notificationService/NotificationServiceImpl.kt
@@ -7,7 +7,9 @@ import com.hedvig.underwriter.model.lastName
 import com.hedvig.underwriter.model.ssnMaybe
 import com.hedvig.underwriter.service.quoteStrategies.QuoteStrategyService
 import com.hedvig.underwriter.serviceIntegration.notificationService.dtos.QuoteCreatedEvent
+import com.hedvig.underwriter.util.logger
 import org.springframework.stereotype.Service
+import java.time.Instant
 
 @Service
 class NotificationServiceImpl(
@@ -17,6 +19,32 @@ class NotificationServiceImpl(
     override fun sendQuoteCreatedEvent(quote: Quote) {
         val event = quoteStrategyService.createQuoteCreatedEvent(quote)
         client.quoteCreated(event)
+    }
+
+    override fun postSignUpdate(quote: Quote) {
+        if (quote.memberId != null) {
+            val isSwitcher = quote.currentInsurer != null
+            try {
+                val map = mapOf(
+                    "partner_code" to quote.attributedTo.name,
+                    "sign_source" to quote.initiatedFrom.name,
+                    "sign_date" to Instant.now(),
+                    "switcher_company" to quote.currentInsurer,
+                    "is_switcher" to isSwitcher
+                )
+                client.post(quote.memberId, map)
+            } catch (ex: Exception) {
+                logger.error("Could not update member with id ${quote.memberId}", ex)
+            }
+        }
+    }
+
+    override fun deleteMember(memberId: String) {
+        try {
+            client.deleteMember(memberId)
+        } catch (e: Exception) {
+            logger.error("Could not delete member: $memberId", e)
+        }
     }
 }
